+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 18.0.2
 
+- Add `allowNull` optional parameter to the `CompareValidator`.
 - Add `oneOf` validator to the list of validators.
 - Add `debounced` async validator that allows to specify a custom debounce time for a single validator.
 - The `Validators.delegateAsync()` function now accepts an optional `debounceTime` parameter, defaulting to 0. This allows immediate execution or custom debouncing for asynchronous validation.